requirements.txt
streamlit
pandas
gspread
google-auth
streamlit-aggrid
google-generativeai
